Use vim-signify for diff symbols
I was able to get vim-gitgutter to work after getting help from the
project owner, but there is still an issue with the sign column briefly
having the wrong color on startup. The ruler also briefly has a wrong
color. It looks like the plugin sets highlight colors, causing issues. I
tried vim-signify, and it works just fine. vim-signify also doesn't
update the signs until I save the file, which is the behavior I want. To
get the same with vim-gitgutter, I'd have to add a file to the
"/after/plugin" directory.